+++ apps/tergite-dashboard/cypress/e2e/admin-requests.cy.ts
@@ -98,6 +98,23 @@ users.forEach((user) => {
       });
 
     isAdmin &&
+      it("does not render number of pending requests when sidebar is collapsed", () => {
+        cy.visit("/");
+        cy.wait("@my-user-info");
+        cy.url().should("equal", "http://127.0.0.1:5173/");
+
+        cy.get("[data-testid='sidebar'] [aria-label='PanelLeftClose']")
+          .parent()
+          .click();
+
+        cy.contains("[data-testid='sidebar'] div", /requests/i).within(() => {
+          cy.contains(".bg-primary", `${pendingUserRequests.length}`).should(
+            "not.exist"
+          );
+        });
+      });
+
+    isAdmin &&
       it("renders the admin requests page when nav item is clicked", () => {
         cy.visit("/");
         cy.wait("@my-user-info");
@@ -335,6 +352,24 @@ users.forEach((user) => {
       });
 
     isAdmin &&
+      it("renders sidebar placeholder by default", () => {
+        cy.viewport(1080, 750);
+        // request summary should not exist
+        cy.get("#request-summary").should("not.exist");
+        cy.contains("#sidebar-placeholder h3", /user request title/i).should(
+          "be.visible"
+        );
+        cy.contains(
+          "#sidebar-placeholder .font-semibold",
+          /request details/i
+        ).should("be.visible");
+        cy.contains(
+          "#sidebar-placeholder .text-muted-foreground",
+          /click any row to show details here/i
+        ).should("be.visible");
+      });
+
+    isAdmin &&
       it("renders user request summary when row is clicked", () => {
         cy.viewport(1080, 750);
         cy.wait(100);
@@ -347,6 +382,9 @@ users.forEach((user) => {
               const requestBody = request.request as AnyFlatRecord;
 
               cy.wrap(obj.el).realClick();
+
+              // sidebar placeholder should not exist
+              cy.get("#sidebar-placeholder").should("not.exist");
 
               cy.contains("#request-summary h3", requestTitle).should(
                 "be.visible"