+++ apps/tergite-dashboard/cypress/e2e/home.cy.ts
@@ -406,35 +406,34 @@ users.forEach((user) => {
             .then(() => {
               cy.contains("[data-cy-job-detail]", /details about this job/i, {
                 timeout: 10000,
-              })
-                .within(() => {
-                  const durationRegex = job.duration_in_secs
-                    ? // FIXME: Not the right regx for ~ '1 days, 2 minutes, 30 seconds' but might work
-                      /Duration:\s?(\d+ (seconds?)|(minutes?)|(hours?)|(days?)|(weeks?)|(months?)|(years?),?)+/i
-                    : /Duration:\s?N\/A/i;
-                  cy.contains(new RegExp(`Job:\\s?${job.job_id}`, "i")).should(
-                    "be.visible"
-                  );
-                  cy.contains(
-                    new RegExp(`Status:\\s?${job.status}`, "i")
-                  ).should("be.visible");
-                  cy.contains(
-                    /Created at:\s?\d+ \w+ \d+, \d+:\d+ ?(AM)|(PM)?/i
-                  ).should("be.visible");
-                  cy.contains(
-                    new RegExp(`Device:\\s?${job.device}`, "i")
-                  ).should("be.visible");
-                  cy.contains(durationRegex).should("be.visible");
-                  job.failure_reason &&
-                    cy
-                      .contains(
-                        new RegExp(`Error:\\s?${job.failure_reason}`, "i")
-                      )
-                      .should("be.visible");
-                })
-                .then(() => {
-                  cy.reload();
-                });
+              }).within(() => {
+                const durationRegex = job.duration_in_secs
+                  ? // FIXME: Not the right regx for ~ '1 days, 2 minutes, 30 seconds' but might work
+                    /Duration:\s?(\d+ (seconds?)|(minutes?)|(hours?)|(days?)|(weeks?)|(months?)|(years?),?)+/i
+                  : /Duration:\s?N\/A/i;
+                cy.contains(new RegExp(`Job:\\s?${job.job_id}`, "i")).should(
+                  "be.visible"
+                );
+                cy.contains(new RegExp(`Status:\\s?${job.status}`, "i")).should(
+                  "be.visible"
+                );
+                cy.contains(
+                  /Created at:\s?\d+ \w+ \d+, \d+:\d+ ?(AM)|(PM)?/i
+                ).should("be.visible");
+                cy.contains(new RegExp(`Device:\\s?${job.device}`, "i")).should(
+                  "be.visible"
+                );
+                cy.contains(durationRegex).should("be.visible");
+                job.failure_reason &&
+                  cy
+                    .contains(
+                      new RegExp(`Error:\\s?${job.failure_reason}`, "i")
+                    )
+                    .should("be.visible");
+
+                // close the drawer
+                cy.get('button[aria-label="X"]').click();
+              });
             });
         });
       }