+++ apps/tergite-dashboard/src/pages/home/index.tsx
@@ -33,7 +33,7 @@ export function Home() {
   );
 
   return (
-    <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8 lg:grid-cols-3 xl:grid-cols-3">
+    <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 lg:grid-cols-3 xl:grid-cols-3">
       <div className="grid auto-rows-max items-start gap-4 lg:col-span-3">
         <div className="grid gap-4 sm:grid-cols-[250px_auto_auto]">
           <Card className="grid-fit-content">