+++ apps/tergite-dashboard/src/pages/projects/index.tsx
@@ -20,7 +20,7 @@ import {
   type Project,
 } from "../../../types";
 import { ProjectsTable } from "./components/project-table";
-import { ProjectSummary } from "./components/project-summary";
+import { ProjectsSidebar } from "./components/projects-sidebar";
 import { Row, RowSelectionState } from "@tanstack/react-table";
 import { useLoaderData } from "react-router-dom";
 
@@ -62,8 +62,8 @@ export function Projects() {
   }, [queryClient, setRowSelection]);
 
   return (
-    <main className="grid flex-1 items-start gap-4 grid-cols-1 p-4 sm:px-6 sm:py-0 md:gap-8 xl:grid-cols-4">
-      <Card className="col-span-1 mt-14  xl:pt-3 xl:col-span-3">
+    <main className="grid flex-1 items-start gap-4 grid-cols-1 p-4 sm:px-6 sm:py-0 xl:grid-cols-4">
+      <Card className="col-span-1 xl:pt-3 xl:col-span-3">
         <CardHeader>
           <CardDescription>Projects</CardDescription>
         </CardHeader>
@@ -79,14 +79,12 @@ export function Projects() {
         </CardContent>
       </Card>
 
-      {projects[selectedProjectIdx] && (
-        <ProjectSummary
-          project={projects[selectedProjectIdx]}
-          className="order-first xl:order-none mt-14 col-span-1"
-          onDelete={handleProjectDelete}
-          canDelete={projects[selectedProjectIdx].admin_id === currentUser.id}
-        />
-      )}
+      <ProjectsSidebar
+        project={projects[selectedProjectIdx]}
+        className="order-first xl:order-none col-span-1"
+        onDelete={handleProjectDelete}
+        canDelete={projects[selectedProjectIdx]?.admin_id === currentUser.id}
+      />
     </main>
   );
 }