+++ apps/tergite-dashboard/src/pages/tokens/index.tsx
@@ -20,7 +20,7 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import { useLoaderData } from "react-router-dom";
 import { type AppState, type ExtendedAppToken, type Project } from "types";
 import { TokensTable } from "./components/tokens-table";
-import { TokenSummary } from "./components/token-summary";
+import { TokensSidebar } from "./components/tokens-sidebar";
 import { Row, RowSelectionState } from "@tanstack/react-table";
 import { Progress } from "@/components/ui/progress";
 
@@ -71,8 +71,8 @@ export function Tokens() {
   }, [currentProject, previousProject, setPreviousProject, setRowSelection]);
 
   return (
-    <main className="grid flex-1 items-start gap-4 grid-cols-1 p-4 sm:px-6 sm:py-0 md:gap-8 xl:grid-cols-4">
-      <Card className="col-span-1 mt-14  xl:pt-3 xl:col-span-3">
+    <main className="grid flex-1 items-start gap-4 grid-cols-1 p-4 sm:px-6 sm:py-0 xl:grid-cols-4">
+      <Card className="col-span-1 xl:pt-3 xl:col-span-3">
         <CardHeader>
           <CardDescription>API tokens for {projectName}</CardDescription>
         </CardHeader>
@@ -89,10 +89,10 @@ export function Tokens() {
         </CardContent>
       </Card>
 
-      {tokens && tokens[selectedTokenIdx] && (
-        <TokenSummary
+      {tokens && (
+        <TokensSidebar
           token={tokens[selectedTokenIdx]}
-          className="order-first xl:order-none mt-14 col-span-1"
+          className="order-first xl:order-none col-span-1"
           onDelete={handleTokenDelete}
         />
       )}